reorganize tools module; break up helper function; remove estimation functions
