+++ src/neuroconv/tools/aws/__init__.py
@@ -1,0 +1,3 @@
+from ._aws import submit_aws_batch_job
+
+__all__ = ["submit_aws_batch_job"]
