+++ tests/test_minimal/test_tools/aws_tools.py
